chore: relocate function file
